+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Basic fuzz testing for `parseInstallationInfo` in `pkg/input/installations`
+
 ## [0.19.0] - 2025-01-27
 
 ### Added